requirements.txt
boto3==1.21.7
botocore==1.24.46
numpy==2.3.0
pandas==2.3.0
pyarrow==20.0.0
python-dotenv==0.19.2
